maple/jvs: fix RFID card reader regression (mushiking, vf4)
The RFID card reader/writer is an MIE device. Fixes regression uncovered
by a429fba21d70a801a9c1e826b5c5283445678f7a.
Also fixes mushikc
[ci skip]
