+++ core/hw/maple/maple_devs.cpp
@@ -1858,386 +1858,6 @@ struct maple_dreamparapara_controller : maple_device
 	}
 };
 
-// Emulates a 838-14245-92 maple to RS232 converter
-// wired to a 838-14243 RFID reader/writer (apparently Saxa HW210)
-struct RFIDReaderWriter : maple_base
-{
-	u32 getStatus() const
-	{
-		// b0: !card switch
-		// b1: state=4	errors?
-		// b2: state=5
-		// b3: state=6
-		// b4: state=7
-		// b5: state=8
-		// b6: card lock
-		// when 0x40 trying to read the card
-		u32 status = 1;
-		if (cardInserted)
-			status &= ~1;
-		if (cardLocked)
-			status |= 0x40;
-		return status;
-	}
-
-	// Surprisingly recipient and sender aren't swapped in the response so we override RawDma for this reason
-	// vf4tuned and mushiking do care
-	u32 RawDma(const u32 *buffer_in, u32 buffer_in_len, u32* buffer_out) override
-	{
-		u32 command=buffer_in[0] &0xFF;
-		//Recipient address
-		u32 reci = (buffer_in[0] >> 8) & 0xFF;
-		//Sender address
-		u32 send = (buffer_in[0] >> 16) & 0xFF;
-		u32 outlen = 0;
-		u32 resp = Dma(command, &buffer_in[1], buffer_in_len - 4, &buffer_out[1], outlen);
-
-		if (reci & 0x20)
-			reci |= maple_GetAttachedDevices(bus_id);
-
-		verify(u8(outlen / 4) * 4 == outlen);
-		buffer_out[0] = (resp << 0 ) | (reci << 8) | (send << 16) | ((outlen / 4) << 24);
-
-		return outlen + 4;
-	}
-
-	u32 dma(u32 cmd) override
-	{
-		switch (cmd)
-		{
-		case MDC_DeviceRequest:
-		case MDC_AllStatusReq:
-			// custom function
-			w32(0x00100000);
-			// function flags
-			w32(0);
-			w32(0);
-			w32(0);
-			//1	area code
-			w8(0xff);				// FF: Worldwide, 01: North America
-			//1	direction
-			w8(0);
-			// Product name (totally made up)
-			wstr("MAPLE/232C CONVERT BD", 30);
-
-			// License (60)
-			wstr(maple_sega_brand, 60);
-
-			// Low-consumption standby current (2)
-			w16(0x0069);	// 10.5 mA
-
-			// Maximum current consumption (2)
-			w16(0x0120);	// 28.8 mA
-
-			return cmd == MDC_DeviceRequest ? MDRS_DeviceStatus : MDRS_DeviceStatusAll;
-
-		case MDCF_GetCondition:
-			w32(0x00100000); // custom function
-
-			return MDRS_DataTransfer;
-
-		case MDC_DeviceReset:
-		case MDC_DeviceKill:
-			return MDRS_DeviceReply;
-
-		// 90	get status
-		//
-		// read test:
-		// d0	?
-		// 91	get last cmd status?
-		// a0	?
-		// 91
-		// a1	read md5 in data
-		//			or data itself if after D4 xx xx xx xx
-		// d4	in=d2 03 aa db
-		// 91
-		//
-		// d9	lock
-		// da	unlock
-		//
-		// write test:
-		// D0
-		// 91
-		// B1 05 06 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 (28 bytes)
-		// 91
-		// B1 0b 06 00 00 00 00 c6 41 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 (28 bytes)
-		// 91
-		// B1 11 06 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 (28 bytes)
-		// 91
-		// B1 17 06 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 (28 bytes)
-		// 91
-		// B1 1d 03 00 00 00 00 00 00 00 00 00 00 00 00 00 00 (16 bytes)
-		// 91
-		// C1: 00 00 00 00
-		// 91
-
-		case 0xD0:
-			d4Seen = false;
-			[[fallthrough]];
-		case 0x90:
-		case 0x91:
-		case 0xA0:
-		case 0xD4:
-		case 0xC1:
-			w32(getStatus());
-			if (cmd == 0xd4)
-				d4Seen = true;
-			return (MapleDeviceRV)0xfe;
-
-		case 0xA1:	// read card data
-			DEBUG_LOG(MAPLE, "RFID card read (data? %d)", d4Seen);
-			w32(getStatus());
-			if (!d4Seen)
-				// serial0 and serial1 only
-				wptr(cardData, 8);
-			else
-				wptr(cardData, sizeof(cardData));
-			return (MapleDeviceRV)0xfe;
-
-		case 0xD9:	// lock card
-			cardLocked = true;
-			w32(getStatus());
-			INFO_LOG(MAPLE, "RFID card %d locked", player_num);
-			return (MapleDeviceRV)0xfe;
-
-		case 0xDA:	// unlock card
-			cardLocked = false;
-			cardInserted = false;
-			w32(getStatus());
-			NOTICE_LOG(MAPLE, "RFID card %d unlocked", player_num);
-			os_notify(i18n::T("Card ejected"), 2000);
-			return (MapleDeviceRV)0xfe;
-
-		case 0xB1:	// write to card
-			{
-				w32(getStatus());
-				u32 offset = r8() * 4;
-				size_t size = r8() * 4;
-				skip(2);
-				DEBUG_LOG(MAPLE, "RFID card write: offset 0x%x len %d", offset, (int)size);
-				rptr(cardData + offset, std::min(size, sizeof(cardData) - offset));
-				saveCard();
-				return (MapleDeviceRV)0xfe;
-			}
-
-		case 0xD1:	// decrement counter
-			{
-				int counter = r8();
-				switch (counter) {
-				case 0x03:
-					counter = 0;
-					break;
-				case 0x0c:
-					counter = 1;
-					break;
-				case 0x30:
-					counter = 2;
-					break;
-				case 0xc0:
-					counter = 3;
-					break;
-				default:
-					WARN_LOG(MAPLE, "Unknown counter selector %x", counter);
-					counter = 0;
-					break;
-				}
-				DEBUG_LOG(MAPLE, "RFID decrement %d", counter);
-				cardData[19 - counter]--;
-				saveCard();
-				w32(getStatus());
-				return (MapleDeviceRV)0xfe;
-			}
-
-		default:
-			INFO_LOG(MAPLE, "RFIDReaderWriter: unknown MAPLE COMMAND %d", cmd);
-			return MDRE_UnknownCmd;
-		}
-	}
-
-	MapleDeviceType get_device_type() override {
-		return MDT_RFIDReaderWriter;
-	}
-
-	void OnSetup() override
-	{
-		memset(cardData, 0, sizeof(cardData));
-		transientData = false;
-	}
-
-	std::string getCardPath() const
-	{
-		int playerNum;
-		if (config::GGPOEnable && !config::ActAsServer)
-			// Always load P1 card with GGPO to be consistent with P1 inputs being used
-			playerNum = 1;
-		else
-			playerNum = player_num + 1;
-		return hostfs::getArcadeFlashPath() + "-p" + std::to_string(playerNum) + ".card";
-	}
-
-	void loadCard()
-	{
-		if (transientData)
-			return;
-		std::string path = getCardPath();
-		FILE *fp = nowide::fopen(path.c_str(), "rb");
-		if (fp == nullptr)
-		{
-			if (settings.content.gameId.substr(0, 8) == "MKG TKOB")
-			{
-				constexpr u8 MUSHIKING_CHIP_DATA[128] = {
-					0x12, 0x34, 0x56, 0x78, // Serial No.0
-					0x31, 0x00, 0x86, 0x07, // Serial No.1
-					0x00, 0x00, 0x00, 0x00, // Key
-					0x04, 0xf6, 0x00, 0xAA, // Extend  Extend  Access  Mode
-					0x23, 0xFF, 0xFF, 0xFF, // Counter4  Counter3  Counter2  Counter1
-					0x00, 0x00, 0x00, 0x00, // User Data (first set date: day bits 0-4, month bits 5-8, year bits 9-... + 2000)
-					0x00, 0x00, 0x00, 0x00, // User Data
-					0x00, 0x00, 0x00, 0x00, // User Data
-					0x00, 0x00, 0x00, 0x00, // User Data
-					0x00, 0x00, 0x00, 0x00, // User Data
-					0x00, 0x00, 0x00, 0x00, // User Data
-					0x23, 0xFF, 0xFF, 0xFF, // User Data (max counters)
-				};
-				memcpy(cardData, MUSHIKING_CHIP_DATA, sizeof(MUSHIKING_CHIP_DATA));
-				for (int i = 0; i < 8; i++)
-					cardData[i] = rand() & 0xff;
-				u32 mask = 0;
-				if (settings.content.gameId == "MKG TKOB 2 JPN VER2.001-"			// mushik2e
-						|| settings.content.gameId == "MKG TKOB 4 JPN VER2.000-")	// mushik4e
-					mask = 0x40;
-				cardData[4] &= ~0xc0;
-				cardData[4] |= mask;
-
-				u32 serial1 = (cardData[4] << 24) | (cardData[5] << 16) | (cardData[6] << 8) | cardData[7];
-				u32 key = ~serial1;
-				key = ((key >> 4) & 0x0f0f0f0f)
-					  | ((key << 4) & 0xf0f0f0f0);
-				cardData[8] = key >> 24;
-				cardData[9] = key >> 16;
-				cardData[10] = key >> 8;
-				cardData[11] = key;
-			}
-			else
-			{
-				constexpr u8 VF4_CARD_DATA[128] = {
-						0x10,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   4,0x6c,   0,   0,
-						   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-						   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-						   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-						   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-						   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-						   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-						   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,0xff
-				};
-				memcpy(cardData, VF4_CARD_DATA, sizeof(VF4_CARD_DATA));
-				// Generate random bytes used by vf4 vanilla to make the card id
-				srand(time(0));
-				cardData[2] = rand() & 0xff;
-				cardData[4] = rand() & 0xff;
-				cardData[5] = rand() & 0xff;
-				cardData[6] = rand() & 0xff;
-				cardData[7] = rand() & 0xff;
-			}
-			INFO_LOG(NAOMI, "Card P%d initialized", player_num + 1);
-		}
-		else
-		{
-			INFO_LOG(NAOMI, "Loading card file from %s", path.c_str());
-			if (fread(cardData, 1, sizeof(cardData), fp) != sizeof(cardData))
-				WARN_LOG(NAOMI, "Truncated or empty card file: %s" ,path.c_str());
-			fclose(fp);
-		}
-	}
-
-	void saveCard() const
-	{
-		if (transientData)
-			return;
-		std::string path = getCardPath();
-		FILE *fp = nowide::fopen(path.c_str(), "wb");
-		if (fp == nullptr)
-		{
-			WARN_LOG(NAOMI, "Can't create card file %s: errno %d", path.c_str(), errno);
-			return;
-		}
-		INFO_LOG(NAOMI, "Saving card file to %s", path.c_str());
-		if (fwrite(cardData, 1, sizeof(cardData), fp) != sizeof(cardData))
-			WARN_LOG(NAOMI, "Truncated write to file: %s", path.c_str());
-		fclose(fp);
-	}
-
-	void serialize(Serializer& ser) const override
-	{
-		maple_device::serialize(ser);
-		ser << cardData;
-		ser << d4Seen;
-		ser << cardInserted;
-		ser << cardLocked;
-	}
-	void deserialize(Deserializer& deser) override
-	{
-		maple_device::deserialize(deser);
-		deser >> cardData;
-		deser >> d4Seen;
-		deser >> cardInserted;
-		deser >> cardLocked;
-	}
-
-	void insertCard()
-	{
-		if (!cardInserted) {
-			cardInserted = true;
-			loadCard();
-		}
-		else if (!cardLocked)
-		{
-			cardInserted = false;
-			if (!transientData)
-				memset(cardData, 0, sizeof(cardData));
-		}
-	}
-
-	const u8 *getCardData() {
-		loadCard();
-		return cardData;
-	}
-
-	void setCardData(u8 *data) {
-		memcpy(cardData, data, sizeof(cardData));
-		transientData = true;
-	}
-
-	u8 cardData[128];
-	bool d4Seen = false;
-	bool cardInserted = false;
-	bool cardLocked = false;
-	bool transientData = false;
-};
-
-void insertRfidCard(int playerNum)
-{
-	std::shared_ptr<maple_device> mapleDev = MapleDevices[1 + playerNum][5];
-	if (mapleDev != nullptr && mapleDev->get_device_type() == MDT_RFIDReaderWriter)
-		std::static_pointer_cast<RFIDReaderWriter>(mapleDev)->insertCard();
-}
-
-void setRfidCardData(int playerNum, u8 *data)
-{
-	std::shared_ptr<maple_device> mapleDev = MapleDevices[1 + playerNum][5];
-	if (mapleDev != nullptr && mapleDev->get_device_type() == MDT_RFIDReaderWriter)
-		std::static_pointer_cast<RFIDReaderWriter>(mapleDev)->setCardData(data);
-}
-
-const u8 *getRfidCardData(int playerNum)
-{
-	std::shared_ptr<maple_device> mapleDev = MapleDevices[1 + playerNum][5];
-	if (mapleDev != nullptr && mapleDev->get_device_type() == MDT_RFIDReaderWriter)
-		return std::static_pointer_cast<RFIDReaderWriter>(mapleDev)->getCardData();
-	else
-		return nullptr;
-}
-
 std::shared_ptr<maple_device> maple_Create(MapleDeviceType type)
 {
 	switch(type)
@@ -2257,7 +1877,7 @@ std::shared_ptr<maple_device> maple_Create(MapleDeviceType type)
 			return std::make_shared<maple_lightgun>();
 		else
 			return std::make_shared<atomiswave_lightgun>();
-	case MDT_NaomiJamma:		return std::make_shared<maple_naomi_jamma>();
+	case MDT_NaomiJamma:		return MIE::Create();
 	case MDT_TwinStick:			return std::make_shared<maple_sega_twinstick>();
 	case MDT_AsciiStick:		return std::make_shared<maple_ascii_stick>();
 	case MDT_MaracasController:	return std::make_shared<maple_maracas_controller>();
@@ -2267,7 +1887,7 @@ std::shared_ptr<maple_device> maple_Create(MapleDeviceType type)
 	case MDT_DenshaDeGoController:	return std::make_shared<maple_densha_controller>();
 	case MDT_SegaControllerXL:	return std::make_shared<FullController>();
 	case MDT_DreamParaParaController:	return std::make_shared<maple_dreamparapara_controller>();
-	case MDT_RFIDReaderWriter:	return std::make_shared<RFIDReaderWriter>();
+	case MDT_RFIDReaderWriter:	return RFIDReaderWriter::Create();
 
 	default:
 		ERROR_LOG(MAPLE, "Invalid device type %d", type);
